+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 Changelog
 =========
 
+### TODO
+- Back button in the import window
+- emission maps: and baking...
+
 ### v 0.1.1
 - Bake button will not activate until the character has been imported and set-up with High-Quality materials.
 - Animation button will not activate until the character has been imported and set-up with either Default or High-Quality materials.
 
-### v 0.1.0 
+### v 0.1.0
 - Initial release.